# Corals/modules/Gateway/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Gateway API
  version: 0.1.0
  description: >
    Issuer API and POS API for the cash-collection payment gateway.
    All money fields are integer centavos (MXN). Every request/response
    carries contract_v: 1. See docs/adapter-contract.md, docs/data-model.md,
    docs/state-machines.md, docs/reference-qr-spec.md for the human contract.
servers:
  - url: /api/v1
    description: Issuer + POS API base path (Sanctum auth, versioned)
security:
  - IssuerAuth: []
...
paths:
  /payment-intents:
    post:
      tags: [Issuer]
      summary: Create a payment intent
      operationId: createPaymentIntent
      security:
        - IssuerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentIntentCreateRequest'
      responses:
        '201':
          description: Payment intent created (ACTIVE — references generated and signed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
        '422':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    IssuerAuth:
      type: http
      scheme: bearer
      description: Sanctum personal access token, scoped abilities, short TTL.
    POSAuth:
      type: http
      scheme: bearer
      description: >
        Per-terminal short-TTL signed JWT (mTLS at the transport layer, enforced
        outside this document). See docs/security-antifraud.md.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Mutating endpoints return the stored response on replay — never a second effect.
      schema:
        type: string
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing/invalid/over-scope credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Authenticated but not authorized for the requested resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Money:
      type: integer
      minimum: 0
      description: Integer centavos (MXN). Never a float.
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    AmountPolicy:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [fixed, variable]
        amount:
          $ref: '#/components/schemas/Money'
        min:
          $ref: '#/components/schemas/Money'
        max:
          $ref: '#/components/schemas/Money'
        allow_partial:
          type: boolean
          default: false
    PaymentIntentCreateRequest:
      type: object
      required: [invoice_id, mid, mode, amount_policy]
      properties:
        invoice_id:
          type: string
        mid:
          type: string
          pattern: '^[0-9]{9}$'
          description: Routes the intent to a merchant owned by the authenticated issuer (403 if not).
        mode:
          type: string
          enum: [one_time, reusable]
        amount_policy:
          $ref: '#/components/schemas/AmountPolicy'
        mapping_strategy:
          type: string
          enum: [deterministic, stored]
          default: deterministic
        expires_at:
          type: string
          format: date-time
        max_payments:
          type: integer
          nullable: true
        overpay_policy:
          type: string
          enum: [reject, accept, accept_and_flag]
          default: reject
        underpay_policy:
          type: string
          enum: [reject, accept, accept_and_flag]
          default: reject
    PaymentReference:
      type: object
      properties:
        human_reference:
          type: string
          description: 12-digit cashier-typable reference (10-digit token + MOD 97-10 check digits).
        barcode_payload:
          type: string
          description: MID(9) + token(10) + checksum(2), Code128-C.
        qr_payload:
          type: string
          description: Signed JSON/CBOR QR payload — see docs/reference-qr-spec.md.
        kid:
          type: string
        expires_at:
          type: string
          format: date-time
        status:
          type: string
    PaymentIntent:
      type: object
      properties:
        public_id:
          type: string
          description: Hashids-encoded id.
        invoice_id:
          type: string
        mode:
          type: string
          enum: [one_time, reusable]
        amount_policy:
          $ref: '#/components/schemas/AmountPolicy'
        state:
          type: string
          enum: [CREATED, ACTIVE, PAID_PENDING_SETTLEMENT, SETTLED, EXPIRED, CANCELED]
        expires_at:
          type: string
          format: date-time
        max_payments:
          type: integer
          nullable: true
        references:
          type: array
          items:
            $ref: '#/components/schemas/PaymentReference'
...